Tabbed plotting window
The widgets that contain the plots are made in separate functions
called makeShellWidget and makeSpectrumWidget. These appear inside
a tabbed widget in the left pane. There is a bug due to which the
spectrum widget is not showing a plot. Before I get to that bug I
am going to do something about the appearance of the plots.