New icons and space for form plus new method of execution
The gui can now be launched in either active or passive mode. In the
active mode it will display enabled icons and menus that allow one to
switch to a widget that will hold the form that can edit the yaml
yaml configuration file. The passive mode is the same old way of
just seeing the results.