User input is now filled into the template dictionary
The input file provided at the beginning is now read and the
config editor is initialized accordingly(there might still be
undiscovered bugs). However, the dictionary
that the config editor outputs is still not ready to execute
reruns. This is because the dictionary picks up all the defaults
for any fields that were not provided even if they are not
mandatory. To fix this there can be two options:

1. If tardis is also putting in default values for the non-mandatory
fields then this can be left as it is.

2. If this is not the case then the output dictionary needs to be
trimmed appropriately.