Added a check to handle older version of matplotlib.
As pointed out by @unoebauer the style-sheets feature was added
to matplotlib in 1.4 version. So I have added a line to check
if the matplotlib version is older than this and skip the line
that applies the stylesheet.

For now a message is printed to ask the users to upgrade their
distribution.