Made setting of QT_API environmental variable mandatory
I had to do this because I found that the import of QT by matplotlib
and IPython go into a conflict all over again if the matplotlib
version is old. So it is better to force them both to import one
version of QT binding that is PySide or PyQt if both are present.
This means that the user has to set the environmetal variable QT_API to
pyside to use PySide or pyqt to use PyQt.