Split the gui into three files
The gui.py has been split into three files datahandler.py,
widgets.py and interface.py.

datahandler declares all the datastructures used by the gui.

widgets handles the design of all the widgets.

the interface instantiates the widgets defined in widgets.py. If
the widget in question requires a data storage object then the
interface also imports the relevant class from the datahandler,
instantiates it and passes it to the widget's constructor.

So the data handling and widget creation are now much less entangled.