Camelcase to PEP8 convention
All methods except the ones that Qt forces us to implement are
now renamed to follow the word1_word2 format instead of the Camelcase.
The methods are also grouped in the class so that the ones that
are Qt forced appear first followed by the custom methods with
comments distinguishing the two.

The class names still follow Camelcase.