Print progress during montecarlo run
This is a quick implementation that prints and updates a status line
containing the how many packets are already finished, how many total
packets are there and the percentage of finished packets.

The format of the statusline can be changed by the gobal macro
STATUS_FORMAT.

The implmentation is exact when running without OMP and an approximation
when activating OMP. This is not garanteed to work bugfree in case the
sheduler for the for-loop changes.
The alternative would be to have a shared variable but that could lead
to a minor performance reduction.

Rename progress tracking variable in cmontecarlo.c

Change STATUS_FORMAT to look nicer