Fix #493
When redirecting the output to a file the escape sequences which are
used to produce the progress updates don't work. This resulted in long
lines in the logile. To avoid this, one can check if the outputstream is
a terminal with isatty().
This functionality was implemented and wrapped in a function defined in
the new header file "io.h". This header will be used in the future to
implement sending logging messages to the python logging framework.

Adding _POSIX_C_SOURCE >=1 is required for fileno to be declared.

Resolves: #493