Change the memory layout to C-Contiguous
Previously accessing two back-to-back elements of a two-dimensional
array resulted in two completely different memmory locations. This resulted in
memory bandwith being the bottleneck for these calculations.

This change converts the arrays to a layout beneficial for our calculations
before passing them to the C extension. The indexing of those arrays is changed
appropriately.
The result of the j_blue_estimator has to be converted back to the format
python expects.
Additionally the j_blue_estimators will only be calculated if the corresponding
array exists. This allows setting line_lists_j_blues to point to Null in order
to skip this part if it is not needed.