Simplify distance2boundary calculation
The calculation that needs to be performed is a simple trigonomic
calculation with two possible results. For performance reason one can
introduce a previous check. For a detailed discussion see #497

This results in the following sheme:
 - For mu > 0 (outward propagation) the next boundary ist the outer
   boundary (case 1)
 - For mu < 0 we have to decide whether we hit the inner boundary or
   not and set the distance accordingly (case 2&3)

Additionally branch prediction optimization was performed in
distance2line. The compiler always assumes if statements to be true,
which means the default branch should always the one the if statement
checks for. This slightly reduces instruction misses.

Resolves: #497