Organize headers in C-Extension
Header files should only contain includes that are necessary for the
compiler to understand the header. Includes needed by the implementation
should be included by the corresponding source.

With this philosophy in mind the includes in all c files were
reorganized.
As part of this reorganization the header `cmontecarlo1.h` was removed
and it's declaration moved to `cmontecarlo.h`

Resolves: #489