Ignore invalid OpenMP nthreads values
If nthreads is not a positive integer omp_set_num_threads won't be
called. The result is that the default number of threads, or if set
OMP_NUM_THREADS is used.
This enables setting nthreads=0 for example to control the number of
threads through the environment (useful for use on clusters)