Return NaN when failing to calculate a distance
When distance2line failed to calculate the distance to the next line the
distance wouldn't change. The result is that the code will continue to
run and it might not be noticed that something failed. Returning NaN
will ensure that the error will be noticed. Until there is proper
handling of errors returned by compute_distance2* it is very risky to
not have values with an unknown state.

Because of this change some tests had to be changed.

test_increment_j_blue_estimator shows how tests should be written
(calculating the expected value during the test according to the known
function). This allows easier parametrization in the future and the
usage of random numbers to test the functions(hypothesis).