Fix issue #552 and #560
The code from move_packet is now mirrored in
move_packet_across_boundary. The current implementation relies on
rpacket.d_boundary being set for the calculation of the estimators.

Some simplifications were made to move_packet_across_boundary to make
the code easier to understand. The underflow of the unsigned integer
next_shell is explicitly used to determine if the packet can cross a
shell boundary or is emitted/reabsorbed.

Add increment_estimators

This separates the propagation of a packet from the incrementing of the
estimators and allows for testing of the estimator increments on top of
being able to have different functions to handle packet propagation.

It is important to note that the estimators should be incremented before
propagating the packet as that changes the value.

Resolves: #552 #560