io/util.py: Change the way YAMLLoader identifies floats
Instead of using yaml's faulty regular expression to implicitly
resolve floats, actually try to convert the property to
float and see if that succeeds.

With this change values like 5e3 will successfully get
parsed as floats, while pyyaml required a format of 5.e+3

While this is more reliable, it is also much slower.