Rewrite AtomData.prepare_atom_data
This commit ensures that all DataFrames that are attributes of
`AtomData` are copies and *not views*. This way the issues with
the C code are avoided (later we need to check how the data is
passed to the C code).

Also, some statements are rewritten to correspond to pandas
guidelines and PEP style.