Move helper functions from atomic to util
Move `atomic_number2symbol` and `symbol2atomic_number` from
the `atomic` module to the `util` module. These functions
are helpers (also used in util and other modules) so they really
belong to the util module.

Also, move `data_path` and `tests_data_path` to the `util` module and
rename them as `get_data_path` and `get_tests_data_path`. The rationale
for the renaming is that pytest confuses the `tests_data_path` with
a unit test.