Add k_packets
A k_packet represents a packet of thermal energy.
Assuming radiative equilibrium k_packets are converted
in situ back to r_packets. This involves selecting the type
of the cooling process (ff- or bf-emission, collisional excitation
or ionization), as well as selecting the individual cooling
process (e.g collisional excitation to level i).