Implement 1st revision of the calling API
Currently the c function expects a storage model, a blackbody
temperature, an array of frequencies, the length of said array, the
source function and the number of integration points. The function
returns a array with the same length as the frequency array. Remember to
free the result to prevent memory leaks.

The python wrapper expects a simulation object, a np.ndarray of
frequencies and the number of integration points.

Additionally the identation of the C code was changed