Add test_spectrum
This concludes the refactoring of TARDISSpectrum.

The syntax for the dummy to_hdf was slightly changed to match the syntax of pandas
to_hdf() methods. That means the 'name' parameter was removed as there
is no reason for the function do decide where it should save to and the
path is now a required argument.