Modify intensity_black_body function
The intensity_black_body function uses numexpr library to speed up the
computation. I've decorated this function using numba with the
intention of optimize it even more. As a result, the numexpr
library is no longer needed in this file

The results were not satisfactory, given that numba's performance
boost relies on calling the function multiple times, which
isn't the case with intensity_black_body