Quantity-ify TARDIS first objective
This PR creates a class DataTable which is capable of storing units of attributes and provide capability to pass these to another object of same class.

Issue Addressed 
---------------------------
Pandas DataFrame isn't capable to store the units of the attributes. The other issue is that this information about units of attributes gets further vanished when we try to copy one dataframe into another. 

Implementation
---------------------------
I have tried to solve this problem by encapsulating the dataframe and units of attributes into a single class. Using this class we can not even store the units of attributes but also pass them into another object of same class by just passing the class members of first object to other.

Testing
---------------------------
I have tested the code by importing the library. I have also added short example about how to use the class in comments of the class.