Fix incorrect docstrings (#1369)
* Fixed to the montecarlo_numba file to compy w/ issue #680

* Fixed the latex equation

* Trying to fix merge conflict

* revert to conf.py

* Remove extra carriage return

* Update tardis/io/decay.py

Co-authored-by: Jaladh Singhal <jaladhsinghal@gmail.com>

* Update decay.py

* Update interaction.py

* Update tardis/plasma/properties/atomic.py

Co-authored-by: Jaladh Singhal <jaladhsinghal@gmail.com>

* Update single_packet_loop.py

* Update atomic.py

Fixed the formatting in regards to @jalahdh-singhal 's comments

* Update numba_interface.py

Changed the numba.array to numba.ndarray

* Update r_packet.py

Changed the r_packet type from RPacket to tardis.montecarlo.montecarlo_numba.r_packet.RPacket

* Update ion_population.py

Changed the Pandas DataFrame to pandas.DataFrame, and Numpy Array to numpy.ndarray. Made sure that the indentation was consistent as well.

* Update atomic.py

Changed to Pandas DataFrame/Series/MultiIndex to be pandas.DataFrame/Series/MultiIndex. Also changed the numpy descriptor to numpy.ndarray, and for now moved the type as a comment.

* fixing merge conflicts

* Changed the returns to fit what type they are

Returns were mostly in : descriptor. Changed them to just be what type, which they are pandas.DataFrame as seen from the inheritance of the class IsotopeAbundances.

* Modification to docstrings and addressing comments

* Update base.py

Removed trailing carriage returns on docstrings

* Update interaction.py

Naming the variables by their full extensions, not just the class that they are.

* Update macro_atom.py

Added the whole path (.NumbaPlasma) to the end of the description so the entire path is present.

* Update r_packet.py

Made the complete path for variables

* Update atomic.py

Updated the descriptions to try and match comments. Fixed the descriptors for the numpy.ndarrays to make them in the same line.

* Update ion_population.py

Fixing the format of docstrings so that the dtype would be same line as the var : type

* Update docs/index.rst

Co-authored-by: Jaladh Singhal <jaladhsinghal@gmail.com>

* Update tardis/plasma/properties/atomic.py

Co-authored-by: Jaladh Singhal <jaladhsinghal@gmail.com>

* Update tardis/io/decay.py

Co-authored-by: Jaladh Singhal <jaladhsinghal@gmail.com>

* Update decay.py

* Update r_packet.py

Co-authored-by: Jaladh Singhal <jaladhsinghal@gmail.com>