Add ability to plot only top contributing elements to SDEC plot (#1506)
* Added nelements functionality

Added ability to specify the number of elements included in the plot. The elements are sorted by their total contribution to the absorption and emission. The top nelements are then given unique colours in the plot. All other elements are shown in silver.

* Changed other element grouping

Changed grouping for other elements. Before it would run through each 'other' element and then all the plot elements. Now all of the 'other' elements are grouped into a single column that is plotted first. Then it goes onto the plot elements. This significantly simplifies the plotting

* Fixed black formatting

* Added comments

* Fixed typos

* Black formatting

* Updated comments and total_luminosity_df

Changed comments based on reviews.

Also changed how the total luminosity df works. Instead of adding 'noint' and 'escatter' columns to the absorption df, then adding to the emission df to get the total, it now drops those columns from the emission df when doing the addition.

* Black formatting

* Update comments

Co-authored-by: Jaladh Singhal <jaladhsinghal@gmail.com>

* Fixed typo

Co-authored-by: Jaladh Singhal <jaladhsinghal@gmail.com>