Add catch for missing elements in SDEC plot (#1517)
* Add catch for missing elements

The elements coloured in the plot are determined via the absorption and emission. It's possible however, that these elements are not the same, e.g. if you specify a limited wavelength range. This could give an error if you try to plot an element not in the emission dataframe, for example. This adds a try statement to catch this case

* Fixed typo

* Black formatting

* Fixed typo