First objective Completed
In a Jupyter notebook, use the requests. post function of Python to download the result as a JSON from this page. You should pass an ENSDF file and the requested format into the data dictionary. You can acquire an ENSDF file from this page by entering “Co56” (Cobalt 56) into the Nucleus input field.