Path.replace bug (#2535)
* Path.replace bug

* made resolve_atom_data_fname to use Path

* Retained old path join way

* Update util.py

Updated using black suggestion

* Can now use data repo

* Update util.py

* Imported Path for conversion.

* explicit type conversion to Path

We have documented this function in certain parts of the documentation to use str. So, I have used explicit type conversion to be compatible with those.

* Check in data repo if `name` with `.h5` is given.

If we have the file in data_repo and the user just gives atom_data name without .h5, we should check in the data repo.